desktop-entry-spec: Add "SingleMainWindow" key
This is an upstreaming of the GNOME-specific X-GNOME-SingleWindow key,
which functions as the same thing, signaling that an app only supports
having a single one of its main window open.

We recently added support for apps with that key in KDE [1] and thought
it would be useful to upstream (with a slightly more accurate name,
since technically single-instance apps can open multiple windows) so
that everyone could benefit from the same thing.

[1] https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/1309