dbus: Also copy busconfig.dtd into place
The canonical system ID (URL) used for the DTD in message bus
configuration is
<http://www.freedesktop.org/standards/dbus/1.0/busconfig.dtd>,
which redirects to
<https://specifications.freedesktop.org/dbus/1.0/busconfig.dtd>,
which is currently a 404.

As with introspect.dtd, it appears that our historical practice has been
to treat 1.0 as a major version here, and incorporate
backwards-compatible changes into it in-place. Use the 1.14.x branch
point (which in practice is newer than the newest change to this file)
as a convenient reference point.

Resolves: https://gitlab.freedesktop.org/dbus/dbus/-/issues/391
Signed-off-by: Simon McVittie <smcv@debian.org>