basedir: Clarify that search path separator should match PATH
The intent of the basedir specification is to add additional environment
variables like `PATH`, to provide other kinds of search paths to apps
(not just searching for executables). Hence the syntax of these
environment variables is like `PATH`, and a colon is used to separate
search paths.

While the XDG basedir spec was never intended to be used on Windows,
it sometimes is, because we have mixed Unix/Windows systems with things
like msys2/mingw/cygwin.

On Windows, the search path separator is a semicolon, because colons are
already used for drive identifiers. So, rather than explicitly say that
the search path separator in the basedir spec should be a colon, say
that it should match the behaviour of `PATH`.

This allows us to use semicolon on Windows (which GLib has been doing
for years) without violating the spec.

I don’t think this can be a backwards compatibility break, as due to
colons being used for drive identifiers on Windows, it’s never really
been possible to use colons as search path separators.

I haven’t explicitly listed semicolons for Windows in the text, as none
of the rest of the spec mentions use on Windows, so I don’t want to
tacitly encourage that kind of use. This change is purely meant to
retcon GLib’s existing behaviour.

See https://gitlab.gnome.org/GNOME/glib/-/issues/3750

Signed-off-by: Philip Withnall <pwithnall@gnome.org>